backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
typing-extensions==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
datetime==5.4
psycopg2-binary==2.9.9